Refactor API extensively
Breaks API down into various focussed subsections relating to clusters
and jobs.
